#7P commit : Clean Code
Make code easier read
- create new login class
