+++ lib/category.dart
@@ -13,8 +13,7 @@ class _State extends State<Category> {
   bool _value4 = false;
   bool _value5 = false;
   bool _value6 = false;
-  bool _value7 = false;
-  bool _value8 = false;
+
 
   //we omitted the brackets '{}' and are using fat arrow '=>' instead, this is dart syntax
   void _value1Changed(bool value) => setState(() => _value1 = value);
@@ -23,8 +22,7 @@ class _State extends State<Category> {
   void _value4Changed(bool value) => setState(() => _value4 = value);
   void _value5Changed(bool value) => setState(() => _value5 = value);
   void _value6Changed(bool value) => setState(() => _value6 = value);
-  void _value7Changed(bool value) => setState(() => _value7 = value);
-  void _value8Changed(bool value) => setState(() => _value8 = value);
+
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
@@ -42,102 +40,67 @@ class _State extends State<Category> {
           new CheckboxListTile(
             value: _value1,
             onChanged: _value1Changed,
-            title: new Text('Stationaries'),
+            title: new Text('Stationery'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('pen,pencil,eraser,notebook,etc.'),
+            subtitle: new Text('Pen, Pencil, Eraser,etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
           new CheckboxListTile(
             value: _value2,
             onChanged: _value2Changed,
-            title: new Text('Clothes'),
+            title: new Text('Clothing'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Student uniform,belt,etc.'),
+            subtitle: new Text('Student Uniform, Belt, etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
           new CheckboxListTile(
             value: _value3,
             onChanged: _value3Changed,
-            title: new Text('Sport Wares'),
+            title: new Text('Sport Equip'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('sport clothes,pants'),
+            subtitle: new Text('Balls, Racquet, Bats, etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
           new CheckboxListTile(
             value: _value4,
             onChanged: _value4Changed,
-            title: new Text('Sport Equipment'),
+            title: new Text('Electronics'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Balls,Racket,Bats,etc.'),
+            subtitle: new Text('Laptop, Tablet, etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
           new CheckboxListTile(
             value: _value5,
             onChanged: _value5Changed,
-            title: new Text('Laptop'),
+            title: new Text('Book'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('ASUS,ACER,Macbook,etc.'),
+            subtitle: new Text('Textbook, Novels, etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
           new CheckboxListTile(
             value: _value6,
             onChanged: _value6Changed,
-            title: new Text('Camera'),
-            controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Mirrorless camera,Dslr,Camera'),
-            secondary: new Icon(Icons.archive),
-            activeColor: Colors.red,
-          ),
-          new CheckboxListTile(
-            value: _value7,
-            onChanged: _value7Changed,
-            title: new Text('Tablets'),
-            controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Apple Ipad,Samsung,galaxy,MIpad,etc.'),
-            secondary: new Icon(Icons.archive),
-            activeColor: Colors.red,
-          ),
-          new CheckboxListTile(
-            value: _value8,
-            onChanged: _value8Changed,
-            title: new Text('Electronic Accessories'),
-            controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Charging cable,powerbank,etc.'),
-            secondary: new Icon(Icons.archive),
-            activeColor: Colors.red,
-          ),
-          new CheckboxListTile(
-            value: _value8,
-            onChanged: _value8Changed,
-            title: new Text('Books'),
-            controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Any Books that you have'),
-            secondary: new Icon(Icons.archive),
-            activeColor: Colors.red,
-          ),
-
-          new CheckboxListTile(
-            value: _value8,
-            onChanged: _value8Changed,
             title: new Text('Others'),
             controlAffinity: ListTileControlAffinity.leading,
-            subtitle: new Text('Other Items that are not listed'),
+            subtitle: new Text('Tools, Lab Equipment, etc.'),
             secondary: new Icon(Icons.archive),
             activeColor: Colors.red,
           ),
-
           new MaterialButton(
             height: 40.0,
             minWidth: 300.0,
             color: Colors.pink[400],
             textColor: Colors.white,
             child: new Text("Finish"),
-            onPressed: () {},
+            onPressed: () {
+              Navigator.popUntil(context, ModalRoute.withName('/signup'));
+              Navigator.of(context).pushReplacementNamed("/login");
+            },
             splashColor: Colors.pink[200],
           ),
         ],
